Update role to remove openshift AI installation (#9289)
* increasing wait time

* Reduced CPU resource for workers

* wqUpdate workload to reflect changes during verion upgrade

THe demo's openshift and openshift AI version is being update to newer version.
since we are now using the openshift AI prebuilt role so rhods relate code is removed.
Starburst version for Helm operator is the same so no change there.

* [rebased] Update workload to reflect changes during verion upgrade

THe demo's openshift and openshift AI version is being update to newer version.
since we are now using the openshift AI prebuilt role so rhods relate code is removed.
Starburst version for Helm operator is the same so no change there.

---------

Co-authored-by: Nate Stephany <nate@redhat.com>