Removed variable, fixed syntax, add more time to task retry (#9300)
* increasing wait time

* Reduced CPU resource for workers

* wqUpdate workload to reflect changes during verion upgrade

THe demo's openshift and openshift AI version is being update to newer version.
since we are now using the openshift AI prebuilt role so rhods relate code is removed.
Starburst version for Helm operator is the same so no change there.

* [rebased] Update workload to reflect changes during verion upgrade

THe demo's openshift and openshift AI version is being update to newer version.
since we are now using the openshift AI prebuilt role so rhods relate code is removed.
Starburst version for Helm operator is the same so no change there.

* Fixed syntax problems with starburst templates

* update the azure aro ai to use the correct service account name (#9281)

set the ai subscription correctly

Signed-off-by: Paul Czarkowski <username.taken@gmail.com>

* Removed unsed variables and reverted variable names

* Removed variable, fixed syntax, add more time to task retry

---------

Signed-off-by: Paul Czarkowski <username.taken@gmail.com>
Co-authored-by: Nate Stephany <nate@redhat.com>
Co-authored-by: Paul Czarkowski <username.taken@gmail.com>