feat: enable gitops bootstrapping to configure app-of-apps sync (#9655)
While this defaults to the same behavior as before, it allows using the
bootstrap app as an app-of-apps (see
[the docs](https://argo-cd.readthedocs.io/en/latest/operator-manual/cluster-bootstrapping/#app-of-apps-pattern)
on the pattern) which is suitable for iterative development. This is
unlikely to be a setting exercised at all in normal deployments, but
there should be little harm in using it even there. The configuration is
expected to be exercised with variables set like this:

```yaml
ocp4_workload_gitops_bootstrap_sync_prune: false
ocp4_workload_gitops_bootstrap_sync_self_heal: false
ocp4_workload_gitops_bootstrap_sync_app_of_apps_allow_sync_config: false
```

The result of these variables is that:

- The bootstrap app will check for updates to the app-of-apps folder, by
  default, every 3 minutes and apply new Argo Applications it finds
  there, removing ones that have been removed there. This helps with,
  for example, renaming applications.
- Individual apps may or may not be configured with automated sync
  including prune and self heal, but if they are configured with prune
  and self heal then they will not be reconciled back to the definitions
  if prune and self heal are manually disabled on that app.

This supports a workflow where bootstrapped apps all enable prune and
self heal by default, and when debugging or developing changes manually
before committing them the behavior can be toggled live in the ArgoCD
interface. When toggled, the bootstrap "app of apps" won't reconcile the
sync policy back to the originally defined value.