fix: enable argocd application template using charts without values (#9657)
Previously, the logic expected _either_ Helm values files _or_ would
use the kustomize-envvar plugin on ArgoCD. There should be a middle
ground where the plugin is not used and values are not provided, which
particularly makes sense for cases where reusable elements of the
workload do not need to be customized in any way, with a consistent
deployment of all ArgoCD-managed workloads while keeping
parameterization inside AgnosticD-managed workloads (like the Showroom
template, for example).

We support this code path with a default parameter that will preserve
previous behavior, while allowing us to toggle off the expectation in
the boostrap workload.